fix(pg): skip unused totals on more last-N readers
Drop the int cast on the skinny COUNT so a huge thread cannot overflow
the total. getContext last-N and summarization paging only consume
messages/hasMore, so they skip the count too. Changelog wording names
the outcome instead of the SQL.

The metadata-filter path still reports the matching-row count: those
rows are already in memory, and using offset+length would make hasMore
always false. Empty includeTotal pages still use the second COUNT;
that is Studio past-the-end, not the agent path.

+++ stores/pg/src/storage/domains/memory/index.ts
@@ -1094,7 +1094,7 @@ export class MemoryPG extends MemoryStorage {
     // occupy $n+1 and $n+2 and are not referenced by the count.
     const rows =
       (await this.#db.readClient.manyOrNone<MessageRowFromDB & { __total?: string | number }>(
-        `SELECT page.*, count_row.__total FROM (${pageSelect}) page CROSS JOIN (SELECT COUNT(*)::int AS "__total" FROM ${tableName} ${whereClause}) count_row`,
+        `SELECT page.*, count_row.__total FROM (${pageSelect}) page CROSS JOIN (SELECT COUNT(*) AS "__total" FROM ${tableName} ${whereClause}) count_row`,
         dataParams,
       )) || [];
